pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gradient-adk"
version = "0.0.1"
description = "Gradient AI Platform Agent Development Kit"
requires-python = ">=3.10"
dependencies = [
	"fastapi>=0.104.0",
	"uvicorn[standard]>=0.24.0",
	"pydantic>=2.5.0",
	"typing-extensions>=4.8.0",
	"typer>=0.12.0",
	"PyYAML>=6.0",
	"aiohttp>=3.8.0",
	"pytest>=8.4.2",
	"pytest-asyncio>=0.21.0",
	"structlog>=24.1.0"
]

[project.scripts]
gradient = "gradient_adk.cli:run"

[tool.setuptools.packages.find]
include = ["gradient_adk*"]